Add files via upload
Establishes a connection to a MongoDB database named 'E-commerce' running on localhost using Mongoose, an ODM (Object Data Modeling) library for MongoDB and Node.js. The connect() function initializes a connection to the specified MongoDB server (localhost) and database ('E-commerce'). This connection allows the Node.js application to interact with the MongoDB database, enabling operations such as saving, querying, updating, and deleting data. Proper error handling and connection management are crucial for ensuring the stability and reliability of the application's interaction with the database.