Update index.js
This updation provide user registration and login with JSON Web Token (JWT) authentication. 
3 main activities done in this update :-
define the key for token 
Add token with login and register API 
Test token with API in postman
the server generates a JWT token using the jsonwebtoken library, which includes the registered user's information as the payload. This token is signed using a secret key and has an expiration time of 2 hours.